Rename `Scheduler` to `SimpleScheduler`
This is a simple rename of the scheduler because we are going to make the
`Scheduler` name a trait in a followup PR.

towards: #140
