+++ cas/BUILD
@@ -11,7 +11,6 @@ rust_binary(
         "//cas/grpc_service:cas_server",
         "//cas/grpc_service:execution_server",
         "//cas/grpc_service:worker_api_server",
-        "//cas/scheduler",
         "//cas/scheduler:default_scheduler_factory",
         "//cas/store",
         "//cas/store:default_store_factory",
