+++ cas/scheduler/tests/simple_scheduler_test.rs
@@ -24,14 +24,14 @@ use action_messages::{
     NameOrPath, SymlinkInfo,
 };
 use common::DigestInfo;
-use config::schedulers::SimpleScheduler;
+use config;
 use error::{make_err, Code, Error, ResultExt};
 use platform_property_manager::{PlatformProperties, PlatformPropertyValue};
 use proto::build::bazel::remote::execution::v2::ExecuteRequest;
 use proto::com::github::allada::turbo_cache::remote_execution::{
     update_for_worker, ConnectionResult, StartExecute, UpdateForWorker,
 };
-use scheduler::{Scheduler, INTERNAL_ERROR_EXIT_CODE};
+use simple_scheduler::{SimpleScheduler, INTERNAL_ERROR_EXIT_CODE};
 use worker::{Worker, WorkerId};
 
 const INSTANCE_NAME: &str = "foobar_instance_name";
@@ -76,7 +76,7 @@ fn make_system_time(add_time: u64) -> SystemTime {
 }
 
 async fn setup_new_worker(
-    scheduler: &Scheduler,
+    scheduler: &SimpleScheduler,
     worker_id: WorkerId,
     props: PlatformProperties,
 ) -> Result<mpsc::UnboundedReceiver<UpdateForWorker>, Error> {
@@ -89,7 +89,7 @@ async fn setup_new_worker(
 }
 
 async fn setup_action(
-    scheduler: &Scheduler,
+    scheduler: &SimpleScheduler,
     action_digest: DigestInfo,
     platform_properties: PlatformProperties,
     insert_timestamp: SystemTime,
@@ -113,7 +113,8 @@ mod scheduler_tests {
     async fn basic_add_action_with_one_worker_test() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x123456789111);
 
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest = DigestInfo::new([99u8; 32], 512);
 
         let mut rx_from_worker = setup_new_worker(&scheduler, WORKER_ID, Default::default()).await?;
@@ -157,8 +158,8 @@ mod scheduler_tests {
     async fn remove_worker_reschedules_multiple_running_job_test() -> Result<(), Error> {
         const WORKER_ID1: WorkerId = WorkerId(0x111111);
         const WORKER_ID2: WorkerId = WorkerId(0x222222);
-        let scheduler = Scheduler::new_with_callback(
-            &SimpleScheduler {
+        let scheduler = SimpleScheduler::new_with_callback(
+            &config::schedulers::SimpleScheduler {
                 worker_timeout_s: WORKER_TIMEOUT_S,
                 ..Default::default()
             },
@@ -293,7 +294,8 @@ mod scheduler_tests {
 
     #[tokio::test]
     async fn worker_should_not_queue_if_properties_dont_match_test() -> Result<(), Error> {
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest = DigestInfo::new([99u8; 32], 512);
         let mut platform_properties = PlatformProperties::default();
         platform_properties
@@ -368,7 +370,8 @@ mod scheduler_tests {
     async fn cacheable_items_join_same_action_queued_test() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x100009);
 
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest = DigestInfo::new([99u8; 32], 512);
 
         let mut expected_action_state = ActionState {
@@ -437,7 +440,8 @@ mod scheduler_tests {
     #[tokio::test]
     async fn worker_disconnects_does_not_schedule_for_execution_test() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x100010);
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest = DigestInfo::new([99u8; 32], 512);
 
         let rx_from_worker = setup_new_worker(&scheduler, WORKER_ID, Default::default()).await?;
@@ -467,8 +471,8 @@ mod scheduler_tests {
     async fn worker_timesout_reschedules_running_job_test() -> Result<(), Error> {
         const WORKER_ID1: WorkerId = WorkerId(0x111111);
         const WORKER_ID2: WorkerId = WorkerId(0x222222);
-        let scheduler = Scheduler::new_with_callback(
-            &SimpleScheduler {
+        let scheduler = SimpleScheduler::new_with_callback(
+            &config::schedulers::SimpleScheduler {
                 worker_timeout_s: WORKER_TIMEOUT_S,
                 ..Default::default()
             },
@@ -556,7 +560,8 @@ mod scheduler_tests {
     async fn update_action_sends_completed_result_to_client_test() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x123456789111);
 
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest = DigestInfo::new([99u8; 32], 512);
 
         let mut rx_from_worker = setup_new_worker(&scheduler, WORKER_ID, Default::default()).await?;
@@ -646,7 +651,8 @@ mod scheduler_tests {
         const GOOD_WORKER_ID: WorkerId = WorkerId(0x123456789111);
         const ROGUE_WORKER_ID: WorkerId = WorkerId(0x987654321);
 
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest = DigestInfo::new([99u8; 32], 512);
 
         let mut rx_from_worker = setup_new_worker(&scheduler, GOOD_WORKER_ID, Default::default()).await?;
@@ -730,7 +736,8 @@ mod scheduler_tests {
     async fn does_not_crash_if_operation_joined_then_relaunched() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x10000f);
 
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest = DigestInfo::new([99u8; 32], 512);
 
         let mut expected_action_state = ActionState {
@@ -834,7 +841,8 @@ mod scheduler_tests {
     async fn run_two_jobs_on_same_worker_with_platform_properties_restrictions() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x123456789111);
 
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest1 = DigestInfo::new([11u8; 32], 512);
         let action_digest2 = DigestInfo::new([99u8; 32], 512);
 
@@ -972,7 +980,8 @@ mod scheduler_tests {
     async fn run_jobs_in_the_order_they_were_queued() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x123456789111);
 
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), || async move {});
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), || async move {});
         let action_digest1 = DigestInfo::new([11u8; 32], 512);
         let action_digest2 = DigestInfo::new([99u8; 32], 512);
 
@@ -1020,8 +1029,8 @@ mod scheduler_tests {
     async fn worker_retries_on_internal_error_and_fails_test() -> Result<(), Error> {
         const WORKER_ID: WorkerId = WorkerId(0x123456789111);
 
-        let scheduler = Scheduler::new_with_callback(
-            &SimpleScheduler {
+        let scheduler = SimpleScheduler::new_with_callback(
+            &config::schedulers::SimpleScheduler {
                 max_job_retries: 2,
                 ..Default::default()
             },
@@ -1147,11 +1156,12 @@ mod scheduler_tests {
         // Since the inner spawn owns this callback, we can use the callback to know if the
         // inner spawn was dropped because our callback would be dropped, which dropps our
         // DropChecker.
-        let scheduler = Scheduler::new_with_callback(&SimpleScheduler::default(), move || {
-            // This will ensure dropping happens if this function is ever dropped.
-            let _drop_checker = drop_checker.clone();
-            async move {}
-        });
+        let scheduler =
+            SimpleScheduler::new_with_callback(&config::schedulers::SimpleScheduler::default(), move || {
+                // This will ensure dropping happens if this function is ever dropped.
+                let _drop_checker = drop_checker.clone();
+                async move {}
+            });
         assert_eq!(dropped.load(Ordering::Relaxed), false);
 
         drop(scheduler);
